Je poursuis les tests, j'ai changé quelques types de columns dans la tables configuration pour les champs de prix et faire aussi en sorte de vérifier si l'emprunteur a atteint son nombre maximal de renouvellements de prêt plutôt que s'il a une demande à son actif afin de lui permettre la fonctionnalité de renouvellement

+++ database/migrations/2024_08_10_211716_create_statistics_table.php
@@ -14,10 +14,10 @@ return new class extends Migration
         Schema::create('statistics', function (Blueprint $table) {
             $table->id();
             $table->string('month');
-            $table->integer('valid_memories_number')->default(value : 0);
-            $table->integer('invalid_memories_number')->default(value : 0);
-            $table->integer('ebooks_number')->default(value : 0);
-            $table->integer('physical_books_number')->default(value : 0);
+            $table->unsignedBigInteger('valid_memories_number')->default(value : 0);
+            $table->unsignedBigInteger('invalid_memories_number')->default(value : 0);
+            $table->unsignedBigInteger('ebooks_number')->default(value : 0);
+            $table->unsignedBigInteger('physical_books_number')->default(value : 0);
             $table->timestamps();
         });
     }
